fix: make the house viewer test draw a real floor
tdd: green

+++ house/room.go
@@ -992,7 +992,7 @@ func (w *RoomEditorPanel) GetViewer() Viewer {
 
 type Viewer interface {
 	gui.Widget
-	Zoom(float64)
+	Zoom(float32)
 	Drag(float64, float64)
 	WindowToBoard(int, int) (float32, float32)
 	BoardToWindow(float32, float32) (int, int)
